Rewrite get_status use case.
